refactor: excercise Table
