+++ src/components/excercise.jsx
@@ -2,7 +2,7 @@ export default function Excercise(
     {problem, children}
 ) {
   return (
-    <div className="space-y-4">
+    <div className="space-y-4 h-dvh">
         <p className="font-primary text-lg">{problem.description}</p>
         <ul className="list-disc pl-5">
             {problem.steps.map((item, index) => (
