+++ src/data/accountingInitialData.js
@@ -1,0 +1,7 @@
+export const initialData = () => [
+    { id: 0, name: 'Bancos', cargo: '', abono: '' },
+    { id: 1, name: 'Pago proveedor Inmobiliaria SA', cargo: '', abono: '' },
+    { id: 2, name: 'Cobro cheque por Maquina Envasadora', cargo: '', abono: '' },
+    { id: 3, name: 'Pago de Zaza SA', cargo: '', abono: '' },
+    { id: 4, name: 'Viáticos a Monterrey', cargo: '', abono: '' }
+]
