+++ src/data/accountingSolutions.js
@@ -1,0 +1,7 @@
+export const solutionData = () =>  [
+    { id: 0, name: 'Bancos', cargo: '850000', abono: '' },
+    { id: 1, name: 'Pago proveedor Inmobiliaria SA', cargo: '', abono: '17700' },
+    { id: 2, name: 'Cobro cheque por Maquina Envasadora', cargo: '120000', abono: '' },
+    { id: 3, name: 'Pago de Zaza SA', cargo: '25000', abono: '' },
+    { id: 4, name: 'Viáticos a Monterrey', cargo: '', abono: '70000' }
+]
